// src/app/@core/models/company.ts
export interface Company{
  id: number;
  name: string;
  email: string;
  hotLine: number;
  dateIncoporation: Date;
  taxCode: string;
  taxDate: Date;
  taxPlace: string;
  headOffice: string;
  numberStaff: string;
  linkWeb: string;
  description: string;
  avatar: string;
  backdropImg: string;
  isDelete: number;
}